// src/gltf-types.d.ts
export type GlTfId = number;

/**
 * Indices of those attributes that deviate from their initialization value.
 */
export interface AccessorSparseIndices {
    /**
     * The index of the bufferView with sparse indices. Referenced bufferView can't have ARRAY_BUFFER or ELEMENT_ARRAY_BUFFER target.
     */
    "bufferView": GlTfId;
    /**
     * The offset relative to the start of the bufferView in bytes. Must be aligned.
     */
    "byteOffset"?: number;
    /**
     * The indices data type.
     */
    "componentType": 5121 | 5123 | 5125 | number;
    "extensions"?: any;
    "extras"?: any;

    [k: string]: any;
}

/**
 * Array of size `accessor.sparse.count` times number of components storing the displaced accessor attributes pointed by `accessor.sparse.indices`.
 */
export interface AccessorSparseValues {
    /**
     * The index of the bufferView with sparse values. Referenced bufferView can't have ARRAY_BUFFER or ELEMENT_ARRAY_BUFFER target.
     */
    "bufferView": GlTfId;
    /**
     * The offset relative to the start of the bufferView in bytes. Must be aligned.
     */
    "byteOffset"?: number;
    "extensions"?: any;
    "extras"?: any;

    [k: string]: any;
}

/**
 * Sparse storage of attributes that deviate from their initialization value.
 */
export interface AccessorSparse {
    /**
     * Number of entries stored in the sparse array.
     */
    "count": number;
    /**
     * Index array of size `count` that points to those accessor attributes that deviate from their initialization value. Indices must strictly increase.
     */
    "indices": AccessorSparseIndices;
    /**
     * Array of size `count` times number of components, storing the displaced accessor attributes pointed by `indices`. Substituted values must have the same `componentType` and number of components as the base accessor.
     */
    "values": AccessorSparseValues;
    "extensions"?: any;
    "extras"?: any;

    [k: string]: any;
}

/**
 * A typed view into a bufferView.  A bufferView contains raw binary data.  An accessor provides a typed view into a bufferView or a subset of a bufferView similar to how WebGL's `vertexAttribPointer()` defines an attribute in a buffer.
 */
export interface Accessor {
    /**
     * The index of the bufferView.
     */
    "bufferView"?: GlTfId;
    /**
     * The offset relative to the start of the bufferView in bytes.
     */
    "byteOffset"?: number;
    /**
     * The datatype of components in the attribute.
     */
    "componentType": 5120 | 5121 | 5122 | 5123 | 5125 | 5126 | number;
    /**
     * Specifies whether integer data values should be normalized.
     */
    "normalized"?: boolean;
    /**
     * The number of attributes referenced by this accessor.
     */
    "count": number;
    /**
     * Specifies if the attribute is a scalar, vector, or matrix.
     */
    "type": "SCALAR" | "VEC2" | "VEC3" | "VEC4" | "MAT2" | "MAT3" | "MAT4" | string;
    /**
     * Maximum value of each component in this attribute.
     */
    "max"?: number[];
    /**
     * Minimum value of each component in this attribute.
     */
    "min"?: number[];
    /**
     * Sparse storage of attributes that deviate from their initialization value.
     */
    "sparse"?: AccessorSparse;
    "name"?: any;
    "extensions"?: any;
    "extras"?: any;

    [k: string]: any;
}